+++ app/controllers/admin_general_controller.rb
@@ -157,16 +157,6 @@ class AdminGeneralController < AdminController
     @tracks_by_type = TrackThing.group('track_type').count
   end
 
-  def debug
-    @admin_current_user = admin_current_user
-    @current_commit = Statistics::General.new.to_h[:alaveteli_git_commit]
-    @current_branch = `git branch | perl -ne 'print $1 if /^\\* (.*)/'`
-    @current_version = ALAVETELI_VERSION
-    repo = `git remote show origin -n | perl -ne 'print $1 if m{Fetch URL: .*github\\.com[:/](.*)\\.git}'`
-    @github_origin = "https://github.com/#{repo}/tree/"
-    @request_env = request.env
-  end
-
   private
 
   def get_events_title